LayoutTests/fast/mediastream/RTCPeerConnection-getConfiguration.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
    <head>
        <script src="../../resources/js-test-pre.js"></script>
    </head>
    <body>
        <script>
            description("Tests RTCPeerConnection get/setConfiguration.");

            var pc = null;
            var configuration = null;
            var iceServers = null;
            var urls = null;

            function checkConfiguration(serversList, iceTransportPolicy, bundlePolicy)
            {
                iceServers = configuration.iceServers
                shouldBe('iceServers.length', serversList.length.toString());
                shouldBeEqualToString('configuration.iceTransportPolicy', iceTransportPolicy);
                shouldBeEqualToString('configuration.bundlePolicy', bundlePolicy);
                for (i = 0; i < iceServers.length; ++i) {
                    shouldBeEqualToString('iceServers[i].username', serversList[i].username);
                    shouldBeEqualToString('iceServers[i].credential', serversList[i].credential);

                    urls = iceServers[i].urls;
                    for (j = 0; j < urls.length; ++j)
                        shouldBeEqualToString('urls[j]', serversList[i].urls[j]);
                }
            }

            pc = new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}]});
            var servers = [{urls:['stun:foo.com'], username:'', credential:''}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'all', 'balanced');

            pc = new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:'stun:bar.com', username:'user2', credential:'x'}]});
            servers = [{urls:['turn:foo.com'], username:'user', credential:'x'}, {urls:['stun:bar.com'], username:'user2', credential:'x'}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'all', 'balanced');

            pc = new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}]});
            servers = [{urls:['turn:foo.com'], username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'all', 'balanced');

            pc = new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransportPolicy:'relay', bundlePolicy:'balanced'});
            var servers = [{urls:['stun:foo.com'], username:'', credential:''}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'relay', 'balanced');

            pc = new webkitRTCPeerConnection({iceServers:[{urls:'stun:foo.com'}], iceTransportPolicy:'relay', bundlePolicy:'max-compat'});
            var servers = [{urls:['stun:foo.com'], username:'', credential:''}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'relay', 'max-compat');

            pc = new webkitRTCPeerConnection({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}], iceTransportPolicy:'relay', bundlePolicy:'max-bundle'});
            servers = [{urls:['turn:foo.com'], username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'relay', 'max-bundle');

            pc.setConfiguration({iceServers:[{urls:'stun:foo.com'}]});
            var servers = [{urls:['stun:foo.com'], username:'', credential:''}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'all', 'balanced');

            pc.setConfiguration({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:'stun:bar.com', username:'user2', credential:'x'}]});
            servers = [{urls:['turn:foo.com'], username:'user', credential:'x'}, {urls:['stun:bar.com'], username:'user2', credential:'x'}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'all', 'balanced');

            pc.setConfiguration({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}]});
            servers = [{urls:['turn:foo.com'], username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'all', 'balanced');

            pc.setConfiguration({iceServers:[{urls:'stun:foo.com'}], iceTransportPolicy:'relay', bundlePolicy:'balanced'});
            var servers = [{urls:['stun:foo.com'], username:'', credential:''}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'relay', 'balanced');

            pc.setConfiguration({iceServers:[{urls:'stun:foo.com'}], iceTransportPolicy:'relay', bundlePolicy:'max-compat'});
            var servers = [{urls:['stun:foo.com'], username:'', credential:''}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'relay', 'max-compat');

            pc.setConfiguration({iceServers:[{urls:'turn:foo.com', username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}], iceTransportPolicy:'relay', bundlePolicy:'max-bundle'});
            servers = [{urls:['turn:foo.com'], username:'user', credential:'x'},{urls:['stun:bar.com', 'turn:foo.com', 'turn:foo2.com'], username:'user2', credential:'x'}];
            configuration = pc.getConfiguration();
            checkConfiguration(servers, 'relay', 'max-bundle');

            finishJSTest();

            window.jsTestIsAsync = true;
            window.successfullyParsed = true;
        </script>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
